requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.29.2
botocore==1.32.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
clip-by-openai==1.1
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
fonttools==4.44.3
fsspec==2023.10.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.25.1
pyarrow==14.0.1
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.0
s3transfer==0.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==1.7.1
torchvision==0.8.2
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
wcwidth==0.2.10
zipp==3.17.0
